move sync and async client into modules
