+++ elastic/examples/basic_async.rs
@@ -57,5 +57,6 @@ fn run() -> Result<(), Box<Error>> {
 }
 
 fn main() {
+    env_logger::init().unwrap();
     run().unwrap();
 }
