+++ elastic/src/client/mod.rs
@@ -472,17 +472,12 @@ For more details see the [`responses`][responses-mod] module.
 pub mod requests;
 pub mod responses;
 
-use uuid::Uuid;
-use futures::Future;
-use futures_cpupool::CpuPool;
-use tokio_core::reactor::Handle;
-use elastic_reqwest::{SyncBody, AsyncBody, SyncElasticClient, AsyncElasticClient};
-use reqwest::Client as SyncHttpClient;
-use reqwest::unstable::async::Client as AsyncHttpClient;
-
 use self::requests::HttpRequest;
-use self::responses::{sync_response, async_response, SyncResponseBuilder, AsyncResponseBuilder};
-use error::{self, Error, Result};
+
+mod sync;
+mod async;
+pub use self::sync::*;
+pub use self::async::*;
 
 pub use elastic_reqwest::RequestParams;
 
@@ -506,367 +501,6 @@ pub trait Sender: private::Sealed + Clone {
     fn send<TRequest, TBody>(&self, req: TRequest, params: &RequestParams) -> Self::Response
         where TRequest: Into<HttpRequest<'static, TBody>>,
               TBody: Into<Self::Body>;
-}
-
-/** A synchronous request sender. */
-#[derive(Clone)]
-pub struct SyncSender {
-    http: SyncHttpClient
-}
-
-impl private::Sealed for SyncSender {}
-
-impl Sender for SyncSender {
-    type Body = SyncBody;
-    type Response = Result<SyncResponseBuilder>;
-
-    fn send<TRequest, TBody>(&self, req: TRequest, params: &RequestParams) -> Self::Response
-        where TRequest: Into<HttpRequest<'static, TBody>>,
-              TBody: Into<Self::Body>
-    {
-        let correlation_id = Uuid::new_v4();
-        let req = req.into();
-
-        info!("Elasticsearch Request: correlation_id: '{}', path: '{}'", correlation_id, req.url.as_ref());
-
-        let res = match self.http.elastic_req(params, req).map_err(|e| error::request(e)) {
-            Ok(res) => {
-                info!("Elasticsearch Response: correlation_id: '{}', status: '{}'", correlation_id, res.status());
-                res
-            },
-            Err(e) => {
-                error!("Elasticsearch Response: correlation_id: '{}', error: '{}'", correlation_id, e);
-                Err(e)?
-            }
-        };
-        
-        Ok(sync_response(res))
-    }
-}
-
-/** A builder for a syncronous client. */
-pub struct SyncClientBuilder {
-    http: Option<SyncHttpClient>,
-    params: RequestParams
-}
-
-impl SyncClientBuilder {
-    /**
-    Create a new client builder.
-
-    By default, a client constructed by this builder will:
-
-    - Send requests to `localhost:9200`
-    - Not use any authentication
-    - Not use TLS
-    */
-    pub fn new() -> Self {
-        SyncClientBuilder {
-            http: None,
-            params: RequestParams::default()
-        }
-    }
-
-    /**
-    Create a new client builder with the given default request parameters.
-    */
-    pub fn from_params(params: RequestParams) -> Self {
-        SyncClientBuilder {
-            http: None,
-            params: params
-        }
-    }
-
-    /**
-    Set the base url. 
-
-    The url must be fully qualified.
-    This method is a convenient alternative to using `params` to specify the `base_url`.
-
-    # Examples
-
-    Specify a base url for the client to send requests to.
-    In this case, the base url is HTTPS, and not on the root path:
-
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .base_url("https://my_es_cluster/some_path");
-    ```
-    */
-    pub fn base_url<I>(mut self, base_url: I) -> Self 
-        where I: Into<String>
-    {
-        self.params = self.params.base_url(base_url);
-
-        self
-    }
-
-    /**
-    Specify default request parameters.
-    
-    # Examples
-    
-    Require all responses use pretty-printing:
-    
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.url_param("pretty", true));
-    ```
-
-    Add an authorization header:
-
-    ```
-    # use elastic::prelude::*;
-    use elastic::http::header::Authorization;
-
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.header(Authorization("let me in".to_owned())));
-    ```
-
-    Specify a base url (prefer the [`base_url`][SyncClientBuilder.base_url] method on `SyncClientBuilder` instead):
-
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.base_url("https://my_es_cluster/some_path"));
-    ```
-
-    [SyncClientBuilder.base_url]: #method.base_url
-    */
-    pub fn params<F>(mut self, builder: F) -> Self
-        where F: Fn(RequestParams) -> RequestParams
-    {
-        self.params = builder(self.params);
-
-        self
-    }
-
-    /** Use the given `reqwest::Client` for sending requests. */
-    pub fn http_client(mut self, client: SyncHttpClient) -> Self {
-        self.http = Some(client);
-
-        self
-    }
-
-    /** 
-    Construct a [`SyncClient`][SyncClient] from this builder. 
-
-    [Client]: struct.Client.html
-    */
-    pub fn build(self) -> Result<SyncClient> {
-        let http = self.http.map(|http| Ok(http))
-                            .unwrap_or(SyncHttpClient::new())
-                            .map_err(|e| error::build(e))?;
-
-        Ok(SyncClient {
-            sender: SyncSender {
-                http: http
-            },
-            params: self.params
-        })
-    }
-}
-
-/** An asynchronous request sender. */
-#[derive(Clone)]
-pub struct AsyncSender {
-    http: AsyncHttpClient,
-    serde_pool: Option<CpuPool>
-}
-
-impl private::Sealed for AsyncSender {}
-
-impl Sender for AsyncSender {
-    type Body = AsyncBody;
-    type Response = Box<Future<Item = AsyncResponseBuilder, Error = Error>>;
-
-    fn send<TRequest, TBody>(&self, req: TRequest, params: &RequestParams) -> Self::Response
-        where TRequest: Into<HttpRequest<'static, TBody>>,
-              TBody: Into<Self::Body>
-    {
-        let serde_pool = self.serde_pool.clone();
-        let correlation_id = Uuid::new_v4();
-        let req = req.into();
-
-        info!("Elasticsearch Request: correlation_id: '{}', path: '{}'", correlation_id, req.url.as_ref());
-
-        let req_future = self.http
-            .elastic_req(params, req)
-            .map_err(move |e| {
-                error!("Elasticsearch Response: correlation_id: '{}', error: '{}'", correlation_id, e);
-                error::request(e)
-            })
-            .map(move |res| {
-                info!("Elasticsearch Response: correlation_id: '{}', status: '{}'", correlation_id, res.status());
-                async_response(res, serde_pool)
-            });
-        
-        Box::new(req_future)
-    }
-}
-
-/** A builder for an asynchronous client. */
-pub struct AsyncClientBuilder {
-    http: Option<AsyncHttpClient>,
-    serde_pool: Option<CpuPool>,
-    params: RequestParams
-}
-
-impl AsyncClientBuilder {
-    /** 
-    Create a new client builder.
-
-    By default, a client constructed by this builder will:
-
-    - Send requests to `localhost:9200`
-    - Not deserialise repsonses on a cpu pool
-    - Not use any authentication
-    - Not use TLS
-    */
-    pub fn new() -> Self {
-        AsyncClientBuilder {
-            http: None,
-            serde_pool: None,
-            params: RequestParams::default()
-        }
-    }
-
-    /**
-    Create a new client builder with the given default request parameters.
-    */
-    pub fn from_params(params: RequestParams) -> Self {
-        AsyncClientBuilder {
-            http: None,
-            serde_pool: None,
-            params: params
-        }
-    }
-
-    /**
-    Set the base url. 
-
-    The url must be fully qualified.
-    This method is a convenient alternative to using `params` to specify the `base_url`.
-
-    # Examples
-
-    Specify a base url for the client to send requests to.
-    In this case, the base url is HTTPS, and not on the root path:
-
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .base_url("https://my_es_cluster/some_path");
-    ```
-    */
-    pub fn base_url<I>(mut self, base_url: I) -> Self 
-        where I: Into<String>
-    {
-        self.params = self.params.base_url(base_url);
-
-        self
-    }
-
-    /**
-    Specify default request parameters.
-    
-    # Examples
-    
-    Require all responses use pretty-printing:
-    
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.url_param("pretty", true));
-    ```
-
-    Add an authorization header:
-
-    ```
-    # use elastic::prelude::*;
-    use elastic::http::header::Authorization;
-
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.header(Authorization("let me in".to_owned())));
-    ```
-
-    Specify a base url (prefer the [`base_url`][SyncClientBuilder.base_url] method on `SyncClientBuilder` instead):
-
-    ```
-    # use elastic::prelude::*;
-    let builder = SyncClientBuilder::new()
-        .params(|p| p.base_url("https://my_es_cluster/some_path"));
-    ```
-
-    [SyncClientBuilder.base_url]: #method.base_url
-    */
-    pub fn params<F>(mut self, builder: F) -> Self
-        where F: Fn(RequestParams) -> RequestParams
-    {
-        self.params = builder(self.params);
-
-        self
-    }
-
-    /** 
-    Use the given `CpuPool` for serialising and deserialising responses.
-
-    If the pool is `None` then responses will be serialised and deserialised on the same thread as the io `Core`.
-
-    # Examples
-
-    Use a cpu pool to serialise and deserialise responses:
-
-    ```
-    # extern crate futures_cpupool;
-    # extern crate elastic;
-    # use elastic::prelude::*;
-    # use futures_cpupool::CpuPool;
-    # fn main() { run().unwrap() }
-    # fn run() -> Result<(), Box<::std::error::Error>> {
-    let pool = CpuPool::new(4);
-
-    let builder = AsyncClientBuilder::new()
-        .serde_pool(pool);
-    # Ok(())
-    # }
-    ```
-    */
-    pub fn serde_pool<P>(mut self, serde_pool: P) -> Self 
-        where P: Into<Option<CpuPool>>
-    {
-        self.serde_pool = serde_pool.into();
-
-        self
-    }
-
-    /** Use the given `reqwest::Client` for sending requests. */
-    pub fn http_client(mut self, client: AsyncHttpClient) -> Self {
-        self.http = Some(client);
-
-        self
-    }
-
-    /** 
-    Construct an [`AsyncClient`][AsyncClient] from this builder. 
-
-    [Client]: struct.Client.html
-    */
-    pub fn build(self, handle: &Handle) -> Result<AsyncClient> {
-        let http = self.http.map(|http| Ok(http))
-                            .unwrap_or(AsyncHttpClient::new(handle))
-                            .map_err(|e| error::build(e))?;
-
-        Ok(AsyncClient {
-            sender: AsyncSender {
-                http: http,
-                serde_pool: self.serde_pool,
-            },
-            params: self.params,
-        })
-    }
 }
 
 /**
@@ -924,59 +558,6 @@ pub struct Client<TSender> {
     sender: TSender,
     params: RequestParams,
 }
-
-/** 
-A synchronous Elasticsearch client.
- 
-# Examples
-
-Create a synchronous `Client` and send a ping request:
-
-```no_run
-# extern crate elastic;
-# use elastic::prelude::*;
-# fn main() { run().unwrap() }
-# fn run() -> Result<(), Box<::std::error::Error>> {
-let client = SyncClientBuilder::new().build()?;
-
-let response = client.request(PingRequest::new())
-                     .send()?
-                     .into_response::<PingResponse>()?;
-# Ok(())
-# }
-```
-*/
-pub type SyncClient = Client<SyncSender>;
-
-/** 
-An asynchronous Elasticsearch client.
-
-# Examples
-
-Create an asynchronous `Client` and send a ping request:
-
-```no_run
-# extern crate futures;
-# extern crate tokio_core;
-# extern crate elastic;
-# use futures::Future;
-# use tokio_core::reactor::Core;
-# use elastic::prelude::*;
-# fn main() { run().unwrap() }
-# fn run() -> Result<(), Box<::std::error::Error>> {
-let mut core = Core::new()?;
-let client = AsyncClientBuilder::new().build(&core.handle())?;
-
-let response_future = client.request(PingRequest::new())
-                            .send()
-                            .and_then(|res| res.into_response::<PingResponse>());
-
-core.run(response_future)?;
-# Ok(())
-# }
-```
-*/
-pub type AsyncClient = Client<AsyncSender>;
 
 #[cfg(test)]
 mod tests {
