+++ elastic/src/client/responses/parse.rs
@@ -87,6 +87,8 @@ See the [`IsOk`][IsOk] trait for more details.
 [IsOk]: trait.IsOk.html
 */
 
+pub(super) use elastic_reqwest::res::parse;
+
 pub use elastic_reqwest::res::parsing::{HttpResponseHead, IsOk, ResponseBody, MaybeOkResponse,
                                         MaybeBufferedResponse, Unbuffered, Buffered};
 pub use elastic_reqwest::res::error::ParseResponseError;
