+++ examples/README.md
@@ -2,4 +2,4 @@
 
 This directory contains some fleshed out examples using `elastic`.
 
-For simple, single-file examples, see [here]().
+For simple, single-file examples, see [here](https://github.com/elastic-rs/elastic/tree/master/elastic/examples).
